Correctly handle scalar types, fix parse of numpy ints
Fixes: https://github.com/pytorch/pytorch/issues/29252

Summary:

There is some inconsist/incorrect code in the handling of
parsing python/numpy scalars that led to issue #29252:

When we allow interpretation of a zero-dim numpy integer value
as a scalar in pytorch, we incorrectly parse the int as a float.

Also fixes the issue described here:
https://github.com/pytorch/pytorch/pull/27628/files#diff-f539198dd366265fb8dc2d661bc5d5bcR1487

Test Plan:

Added a unit test based on the example given in the issue.

ghstack-source-id: 027bca5a0770bc5daa805bf18e40c1891f49b299
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30486